made tests pass in both qa and demo (extracted test helpers, removed unfinished pusher tests, modified rest session tests)
